+++ src/features/account/model/useLogIn.ts
@@ -1,0 +1,7 @@
+import { useForm } from 'react-hook-form';
+
+export const useLogIn = () => {
+  const form = useForm();
+
+  return { form };
+};
