+++ src/features/account/ui/LoginForm.tsx
@@ -1,0 +1,55 @@
+'use client';
+
+import {
+  Form,
+  FormField,
+  FormItem,
+  FormControl,
+  FormMessage,
+} from '@/shared/ui/ui/form';
+import { Input } from '@/shared/ui/ui/input';
+import { Button } from '@/shared/ui/ui/button';
+import { useLogIn } from '../model/useLogin';
+
+const LoginForm = () => {
+  const { form } = useLogIn();
+
+  return (
+    <Form {...form}>
+      <form
+        className="space-y-4 mt-2 w-full"
+        onSubmit={form.handleSubmit((data) => console.log(data))}
+      >
+        <FormField
+          control={form.control}
+          name="username"
+          render={({ field }) => (
+            <FormItem>
+              <FormControl>
+                <Input placeholder="이메일" type="email" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="password"
+          render={({ field }) => (
+            <FormItem>
+              <FormControl>
+                <Input type="password" placeholder="비밀번호" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <Button type="submit" className="w-full">
+          로그인
+        </Button>
+      </form>
+    </Form>
+  );
+};
+
+export default LoginForm;
